requirements.txt
Flask>=2.0
pymongo>=4.0
python-dotenv>=1.0
python -m venv venv
source venv/bin/activate   # Windows: venv\Scripts\activate
pip install -r requirements.txt
